Server/requirements.txt
paho-mqtt==1.6.1
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0 